refactor(solidlsp): Phase 4 — extract DocumentSymbolRepository + symbol model
- Move raw+unified caches, load/save/version/fingerprint, request_document_symbols orchestration, _request_document_symbols hook, and conversion helpers into:
  - src/solidlsp/ls_symbol_cache.py (DocumentSymbolRepository)
  - src/solidlsp/ls_symbol_model.py (DocumentSymbols, SymbolBody, SymbolBodyFactory, RawDocumentSymbol, convert_raw_document_symbols_to_unified)
- Re-export DocumentSymbols/SymbolBody/SymbolBodyFactory/RawDocumentSymbol/ReferenceInSymbol from solidlsp.ls for pickle/backcompat.
- Bump RAW_DOCUMENT_SYMBOLS_CACHE_VERSION (1→2) and DOCUMENT_SYMBOL_CACHE_VERSION (4→5).
- Facade retains save_cache(), _normalize_symbol_name, _document_symbols_cache_fingerprint, _request_document_symbols as hooks; delegates to repository via narrow callbacks.
- Add facade property shims for _document_symbols_cache/_raw_document_symbols_cache so tests and direct accesses continue to work.
- All public request_* signatures and adapter override hook names unchanged.
